perf: share the stack in the evm context
