+++ crates/handler/src/handler.rs
@@ -6,6 +6,7 @@ use crate::{
 use context::result::{ExecutionResult, FromStringError};
 use context::LocalContextTr;
 use context_interface::context::ContextError;
+use context_interface::local::{FrameStack, FrameToken, OutFrame};
 use context_interface::ContextTr;
 use context_interface::{
     result::{HaltReasonTr, InvalidHeader, InvalidTransaction},
@@ -14,7 +15,6 @@ use context_interface::{
 use interpreter::{FrameInput, Gas, InitialAndFloorGas};
 use primitives::U256;
 use state::EvmState;
-use std::vec::Vec;
 
 pub trait EvmTrError<EVM: EvmTr>:
     From<InvalidTransaction>
@@ -409,16 +409,17 @@ pub trait Handler {
         evm: &mut Self::Evm,
         frame: Self::Frame,
     ) -> Result<FrameResult, Self::Error> {
-        let mut frame_stack = FrameStack::<Self::Frame>::new(frame);
+        let frame_stack = frame_stack::<Self::Frame>;
+        frame_stack(evm).init(frame);
         loop {
-            let (frame, new_frame) = frame_stack.get();
+            let (frame, new_frame) = frame_stack(evm).get();
             let call_or_result = self.frame_call(frame, evm)?;
 
             let result = match call_or_result {
                 ItemOrResult::Item(init) => {
                     match self.frame_init(frame, new_frame, evm, init)? {
                         ItemOrResult::Item(token) => {
-                            frame_stack.push(token);
+                            frame_stack(evm).push(token);
                             continue;
                         }
                         // Do not pop the frame since no new frame was created
@@ -427,15 +428,15 @@ pub trait Handler {
                 }
                 ItemOrResult::Result(result) => {
                     // Remove the frame that returned the result
-                    if frame_stack.index() == 0 {
+                    if frame_stack(evm).index() == 0 {
                         return Ok(result);
                     }
-                    frame_stack.pop();
+                    frame_stack(evm).pop();
                     result
                 }
             };
 
-            self.frame_return_result(frame_stack.get().0, evm, result)?;
+            self.frame_return_result(frame_stack(evm).get().0, evm, result)?;
         }
     }
 
@@ -529,150 +530,8 @@ pub trait Handler {
     }
 }
 
-/// Non-empty, item re-using Vec.
-pub struct FrameStack<T> {
-    stack: Vec<T>,
-    index: usize,
-}
-
-impl<T> FrameStack<T> {
-    /// Creates a new stack with the first item.
-    #[inline]
-    pub fn new(first: T) -> Self {
-        let mut this = Self {
-            stack: Vec::new(),
-            index: 0,
-        };
-        this.stack.push(first);
-        this
-    }
-
-    /// Returns the current index of the stack.
-    #[inline]
-    pub fn index(&self) -> usize {
-        self.index
-    }
-
-    /// Increments the index.
-    #[inline]
-    pub fn push(&mut self, token: FrameToken) {
-        token.assert();
-        if self.index + 1 == self.stack.len() {
-            unsafe { self.stack.set_len(self.stack.len() + 1) };
-            self.stack.reserve(1);
-        }
-        self.index += 1;
-    }
-
-    /// Decrements the index.
-    #[inline]
-    pub fn pop(&mut self) {
-        self.index -= 1;
-    }
-
-    /// Returns the current item.
-    #[inline]
-    pub fn get(&mut self) -> (&mut T, OutFrame<'_, T>) {
-        unsafe {
-            let ptr = self.stack.as_mut_ptr().add(self.index);
-            (
-                &mut *ptr,
-                OutFrame::new_maybe_uninit(ptr.add(1), self.index + 1 < self.stack.len()),
-            )
-        }
-    }
-}
-
-/// A potentially initialized frame. Used when initializing a new frame in the main loop.
-pub struct OutFrame<'a, T> {
-    ptr: *mut T,
-    init: bool,
-    lt: core::marker::PhantomData<&'a mut T>,
-}
-
-impl<'a, T> OutFrame<'a, T> {
-    pub fn new_init(slot: &'a mut T) -> Self {
-        unsafe { Self::new_maybe_uninit(slot, true) }
-    }
-
-    pub fn new_uninit(slot: &'a mut core::mem::MaybeUninit<T>) -> Self {
-        unsafe { Self::new_maybe_uninit(slot.as_mut_ptr(), false) }
-    }
-
-    pub unsafe fn new_maybe_uninit(ptr: *mut T, init: bool) -> Self {
-        Self {
-            ptr,
-            init,
-            lt: Default::default(),
-        }
-    }
-
-    pub fn get(&mut self, f: impl FnOnce() -> T) -> &mut T {
-        if !self.init {
-            self.do_init(f);
-        }
-        unsafe { &mut *self.ptr }
-    }
-
-    #[cold]
-    fn do_init(&mut self, f: impl FnOnce() -> T) {
-        unsafe {
-            self.init = true;
-            self.ptr.write(f());
-        }
-    }
-
-    pub unsafe fn get_unchecked(&mut self) -> &mut T {
-        debug_assert!(self.init, "OutFrame must be initialized before use");
-        unsafe { &mut *self.ptr }
-    }
-
-    pub fn consume(self) -> FrameToken {
-        FrameToken(self.init)
-    }
-}
-
-/// Used to guarantee that a frame is initialized before use.
-pub struct FrameToken(bool);
-
-impl FrameToken {
-    #[cfg_attr(debug_assertions, track_caller)]
-    pub fn assert(&self) {
-        assert!(self.0, "FrameToken must be initialized before use");
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn frame_stack() {
-        let mut stack = FrameStack::new(1);
-
-        assert_eq!(stack.index(), 0);
-        assert_eq!(stack.stack.len(), 1);
-
-        let (a, mut b) = stack.get();
-        assert_eq!(a, &mut 1);
-        assert!(!b.init);
-        assert_eq!(b.get(|| 2), &mut 2);
-        let token = b.consume();
-        stack.push(token);
-
-        assert_eq!(stack.index(), 1);
-        assert_eq!(stack.stack.len(), 2);
-        let (a, b) = stack.get();
-        assert_eq!(a, &mut 2);
-        assert!(!b.init);
-
-        stack.pop();
-
-        assert_eq!(stack.index(), 0);
-        assert_eq!(stack.stack.len(), 2);
-        let (a, mut b) = stack.get();
-        assert_eq!(a, &mut 1);
-        assert!(b.init);
-        assert_eq!(unsafe { b.get_unchecked() }, &mut 2);
-    }
+#[inline]
+fn frame_stack<'a, F: Frame<Evm: EvmTr>>(evm: &mut F::Evm) -> &'a mut FrameStack<F> {
+    let f = evm.ctx_mut().local_mut().frame_stack();
+    unsafe { core::mem::transmute::<&mut FrameStack<u128>, &mut FrameStack<F>>(f) }
 }
